+++ src/main/java/com/tajutechgh/todobackend/exception/ResourceNotFoundException.java
@@ -8,9 +8,9 @@ public class ResourceNotFoundException extends RuntimeException{
 
     private String resourceName;
     private String fieldName;
-    private Long fieldValue;
+    private Integer fieldValue;
 
-    public ResourceNotFoundException(String resourceName, String fieldName, Long fieldValue){
+    public ResourceNotFoundException(String resourceName, String fieldName, Integer fieldValue){
         super(String.format("%s not found with %s : '%s'", resourceName, fieldName, fieldValue));
         this.resourceName = resourceName;
         this.fieldName = fieldName;
